Module FPauth_core

FPauth_core is a library for easy yet customizable authentication in Dream web-applications

module Static : sig ... end

Static is a module containg static type definitions, which are not dependent on Auth_sign.MODEL

module Variables : sig ... end

VARIABLES is a module which inits and holds field variables for authentication

module Session_manager : sig ... end

Session_manager contains

module Authenticator : sig ... end

Authenticator is module which provides functions both for authentication and logout

module Router : sig ... end

Router is a module which contains handlers for authentication and creates routes for them.

module Make_Auth (M : Auth_sign.MODEL) : sig ... end

Make_Auth creates a module based on Auth_sign.MODEL. Provides local variables, middlewares and authenticator to run authentication strategies.

module Auth_sign : sig ... end

Auth_sign is a module containig signatures for modules which can be implemented and integrated from outside the lib, as well as signatures for some inner modules.