Module FPauth_core.Auth_sign

Auth_sign is a module containig signatures for modules which can be implemented and integrated from outside the lib, as well as signatures for some inner modules.

module type MODEL = sig ... end

MODEL is a signature for modules handling authenticable entities

module type SESSIONMANAGER = sig ... end

SESSIONMANAGER is a signature for a functor producing modules for controlling sessions and authentications of entities with type MODEL.t

module type STRATEGY = sig ... end

STRATEGY is a module which contains functions for entity authentications in a certain method, as well as supporting routes and functions

module type AUTHENTICATOR = sig ... end

AUTHENTICATOR is a signature for a functor to create authenticators of various entities over various strategies (See STRATEGY)

module type VARIABLES = sig ... end

VARIABLES is a module containing field variables based on MODEL

module type RESPONSES = sig ... end

RESPONSES is a module which defines how the library should represent some basic events

module type ROUTER = sig ... end

ROUTER is a module which contains handlers for authentication and creates routes for them.