Module type Auth_sign.RESPONSES

RESPONSES is a module which defines how the library should represent some basic events

val login_successful : Dream.request -> Dream.response Dream.promise

login_successful is triggered when authentication has been successful

val login_error : Dream.request -> Dream.response Dream.promise

login_error is triggered if there was any kind of failure during authentication

val logout : Dream.request -> Dream.response Dream.promise

logout is triggered after authentication has been reset