Module type Auth_sign.STRATEGY

STRATEGY is a module which contains functions for entity authentications in a certain method, as well as supporting routes and functions

type entity

type entity is a type of authenticatable entity equal to MODEL.t

val call : Dream.request -> entity -> entity Static.StratResult.t Lwt.t

call is a core function of a strategy. It determines ways of authenticating an entity

val routes : Dream.route

routes defines some additional routes to handlers of a strategy if they are needed. Can contain multiple routes using Dream.scope

val name : Base.string

name is a name of the STRATEGY. Used to define wether the strat can be applied to a certain entity.