Module type Auth_sign.VARIABLES

VARIABLES is a module containing field variables based on MODEL

type entity

type entity is a type of authenticatable entity equal to MODEL.t

val authenticated : Base.bool Dream.field

authenticated is a variable valid for a single request, indicates if authentication has been previously completed. Should be set in SESSIONMANAGER.auth_setup

val current_user : entity Dream.field

current_user is a variable valid for a single request, holds an authenticated entity from a session. Should be set in SESSIONMANAGER.auth_setup

val auth_error : Base.Error.t Dream.field

auth_error is a field with error which occured during any stage of authentication