Module Authenticator.Make

Make creates an instance of Auth_sign.AUTHENTICATOR for a given model and variables

Parameters

module M : Auth_sign.MODEL

module V : Auth_sign.VARIABLES with type entity = M.t

Signature

type entity = M.t

type entity is a type of authenticatable entity equal to MODEL.t

type strategy = (module Auth_sign.STRATEGY with type entity = entity)

strategy is a function that authenticates an entity from a request.

val authenticate : strategy Base.list -> Dream.request -> Static.AuthResult.t Dream.promise

authenticate runs several authentication strategies for a request and defines, whether overall authentication was successful or not

val logout : Dream.request -> Base.unit Lwt.t

logout invalidates session, which resets authentication status to false