Module FPauth_core.Router

Router is a module which contains handlers for authentication and creates routes for them.

module Make (M : Auth_sign.MODEL) (A : Auth_sign.AUTHENTICATOR with type entity = M.t) (V : Auth_sign.VARIABLES) : Auth_sign.ROUTER with type entity = M.t

Make creates an instance of Router with all its dependencies