Parameter Make.2-V

type entity = M.t

type entity is a type of authenticatable entity equal to MODEL.t

val authenticated : Base.bool Dream.field

authenticated is a variable valid for a single request, indicates if authentication has been previously completed. Should be set in SESSIONMANAGER.auth_setup

val current_user : entity Dream.field

current_user is a variable valid for a single request, holds an authenticated entity from a session. Should be set in SESSIONMANAGER.auth_setup

val auth_error : Base.Error.t Dream.field

auth_error is a field with error which occured during any stage of authentication

val update_current_user : entity -> Dream.request -> Base.unit Dream.promise

update_current_user user request updates current_user and session.