Module FPauth_core.Static

Static is a module containg static type definitions, which are not dependent on Auth_sign.MODEL

Static is a module containing all FPauth features, which are not dependant on FPauth.Auth_sign.MODEL

module StratResult : sig ... end

StratResult defines results of strategies, as well as some helpful functions.

module AuthResult : sig ... end

AuthResult is a result of full authentication process. Similar to StratResult, but doesn't have some types which are meaningful only for strategies. Authenticated and Rescue loose content as it is stored in Dream.field by the end of authentication

module Params : sig ... end

Params stores params of a request, either all or only required for authentication