Module Static.AuthResult

AuthResult is a result of full authentication process. Similar to StratResult, but doesn't have some types which are meaningful only for strategies. Authenticated and Rescue loose content as it is stored in Dream.field by the end of authentication

type t =

| Authenticated	(*
Entity has been authenticated successfully.
*)

| Rescue	(*
Authentication must be stopped immediately with an error.
*)

| Redirect of Dream.response Lwt.t	(*
User should be redirected in accordance with response. response promise is meant to be created by Dream.redirect
*)