Module FPauth_strategies

module Password : sig ... end

Password is a simple authentication strategy which verifies identity via provided in params password.

module TOTP : sig ... end

TOTP is a time-based One-Time Password strategy. User's identity is verified via a password which is limited for a limited time only.