Module Password.Make

Make creates a strategy for a provided model.

Parameters

module M : MODEL

Signature

type entity = M.t

val call : Dream.request -> entity -> entity FPauth_core.Static.StratResult.t Lwt.t

call is a main function of a strategy, which authenticates the user by "password" param. The param is verified against a hashed password with Argon2.

val routes : Dream.route

This strategy has no routes and returns Dream.no_route

val name : string

See Password.name