Module type Password.MODEL

MODEL contains requirements for user model in order to use the strategy

type t

val encrypted_password : t -> string option

encrypted_password is a string of hashed password, against which a given password will be verified. Argon2 is used for verification.