Module TOTP.Make_HTML_Responses

This functor creates module with RESPONSES in HTML format

Parameters

module S : HTML_settings

Signature

val response_error : Dream.request -> Base.Error.t -> Dream.response Lwt.t

This response is used to display all kinds of errors

val response_secret : Dream.request -> string -> Dream.response Lwt.t

This response is used during TOTP setup. During this step users are provided with a secret, which he needs to put in his OTP-generator.

val response_enabled : Dream.request -> Dream.response Lwt.t

This response informs users that their TOTP has been enabled