Module type TOTP.RESPONSES

RESPONSES contains data representations for certain events

val response_error : Base.Error.t -> Dream.response Lwt.t

This response is used to display all kinds of errors

val response_secret : string -> Dream.response Lwt.t

This response is used during TOTP setup. During this step users are provided with a secret, which he needs to put in his OTP-generator.

val response_enabled : Dream.response Lwt.t

This response informs users that their TOTP has been enabled