Module FPauth

FPauth is a library for easy yet customizable authentication in Dream web-applications

module Static = FPauth_core.Static

Static is a module containg static type definitions, which are not dependent on Auth_sign.MODEL

module Make_Auth (M : FPauth_core.Auth_sign.MODEL) : sig ... end

Make_Auth creates a module based on Auth_sign.MODEL. Provides local variables, middlewares and authenticator to run authentication strategies.

module Auth_sign = FPauth_core.Auth_sign

Auth_sign is a module containig signatures for modules which can be implemented and integrated from outside the lib, as well as signatures for some inner modules.

module Strategies = FPauth_strategies

Strategies contain two default strategies: FPauth_strategies.Password and FPauth_strategies.Otp

module Responses = FPauth_responses

Responses contain some default responses to basic authentication events