Module FPauth.Make_Auth

Make_Auth creates a module based on Auth_sign.MODEL. Provides local variables, middlewares and authenticator to run authentication strategies.

Parameters

module M : FPauth_core.Auth_sign.MODEL

Signature

module Variables : FPauth_core.Auth_sign.VARIABLES with type entity = M.t

Variables contains types, functions and local variables based on Auth_sign.MODEL.

module Session_manager : FPauth_core.Auth_sign.SESSIONMANAGER with type entity = M.t

SessionManager is a module that sets local variables from session for every request via Auth_sign.SESSIONMANAGER.auth_setup middleware

module Authenticator : FPauth_core.Auth_sign.AUTHENTICATOR with type entity = M.t

Authenticator contains functions for running FPauth.Auth_sign.STRATEGY list and performing logouts

module Router : FPauth_core.Auth_sign.ROUTER with type entity = M.t

Router creates routes, needed for authentication. Contains some basic handlers and joins them with routes from strategies.